+++ R/ColumnLinkedMatrix.R
@@ -315,13 +315,19 @@ setMethod("initialize", signature(.Object = "ColumnLinkedMatrix"), function(.Obj
     if (length(nodes) == 0) {
         nodes[[1]] <- matrix()
     } else {
-        # Detect non-matrix objects by checking dimensions
+        # Stop if matrices are not matrix-like
         if (any(sapply(nodes, function(x) length(dim(x)) != 2))) {
             stop("arguments need to be matrix-like")
         }
-        # Detect matrices that do not match in dimensions
+        # Stop if dimensions of matrices do not match
         if (length(unique(sapply(nodes, nrow))) != 1) {
             stop("arguments need the same number of rows")
+        }
+        # Warn if rownames of matrices do not match
+        names <- lapply(nodes, rownames)
+        names <- names[!sapply(names, is.null)]
+        if (length(names) > 1 && !all(duplicated(names) | duplicated(names, fromLast = TRUE))) {
+            warning("row names of matrix-like objects do not match")
         }
     }
     .Object <- callNextMethod(.Object, nodes)